pyproject.toml
[project]
name = "firestore-schema-dump"
version = "0.1.0"
description = "A robust tool for exploring and documenting Firestore database schemas"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "google-cloud-firestore>=2.20.2",
    "python-dotenv>=1.0.0",
    "rich>=14.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-timeout>=2.1.0",
    "pytest-cov>=4.1.0",
]

[build-system]
requires = ["setuptools>=42"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
testpaths = ["tests"]
timeout = 30

[tool.coverage.run]
source = ["main.py"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if __name__ == .__main__.:",
]